backend/requirements.txt
anyio==3.6.1
asgiref==3.5.1
bcrypt==3.2.2
black==22.3.0
cffi==1.15.0
click==8.1.3
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.77.1
h11==0.13.0
idna==3.3
motor==3.0.0
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.5.2
pycparser==2.21
pydantic==1.9.0
PyJWT==2.4.0
pymongo==4.1.1
python-decouple==3.6
sniffio==1.2.0
starlette==0.19.1
tomli==2.0.1
typing-extensions==4.2.0
uvicorn==0.17.6
